Fix bug in re-matching of affected indices.
It is proper to use the max range of all processes instead of that of
picked process to get the superset indices which would be rematched.

Because the ranges of processes may be different, if the short process
performed and the range of that process is used, then the superset would
miss the sites which is listed in processes with longer range.